some ui refactorings + remove viewBox from SVG + move the aria-title into the SVG as <title>
