+++ src/renderer/components/layout/Footer.tsx
@@ -10,7 +10,7 @@ const closeApp = () => {
     ipcRenderer.send('close-application')
 }
 
-export const Footer = () => (
+export const Footer: React.FunctionComponent = () => (
     <Flex flexWrap='wrap' justifyContent='flex-end' variant='barGradient' sx={{ overflow: 'hidden', borderTop: '1px solid', borderColor: 'shadow' }}>
         <RouterLink to='/config' title='Settings'>
             <Box p={2} color='gray' sx={{ ':hover': { color: 'black' } }}>
