+++ src/renderer/components/layout/Header.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import { Box, Heading } from 'rebass'
 
-export const Header = () => (
+export const Header: React.FunctionComponent = () => (
     <Box p={2} variant='barGradient' sx={{ borderBottom: '1px solid', overflow: 'hidden', borderColor: 'shadow' }}>
         <Heading textAlign='center' fontWeight='bold' fontSize={2}>
             Merge Requests
